add sound playing, saving selections and a lot of clean up of visualization functions. (#45, #49)
