+++ bacpipe/embedding_evaluation/visualization/visualize.py
@@ -3,19 +3,16 @@ import json
 import matplotlib.pyplot as plt
 from matplotlib.figure import Figure
 import numpy as np
-from pathlib import Path
 
 import bacpipe.embedding_evaluation.label_embeddings as le
 import matplotlib
 import seaborn as sns
-from matplotlib.colors import rgb2hex
 
 import logging
 
 logger = logging.getLogger(__name__)
 
 
-SPECTROGRAM_PLOT_HEIGHT = 300
 
 matplotlib.rcParams.update(
     {
@@ -28,1001 +25,6 @@ matplotlib.rcParams.update(
         "ytick.labelsize": 10,
     }
 )
-
-
-def darken_hex_color_bitwise(hex_color):
-    """
-    Darkens a hex color using the bitwise operation: (color & 0xfefefe) >> 1.
-
-    Parameters:
-        hex_color (str): The hex color string (e.g., '#1f77b4').
-
-    Returns:
-        str: The darkened hex color.
-    """
-    # Remove '#' and convert hex color to an integer
-    color_int = int(hex_color.lstrip("#"), 16)
-
-    # Apply the bitwise operation to darken the color
-    darkened_color_int = (color_int & 0xFEFEFE) >> 1
-
-    # Convert back to a hex string and return with leading '#'
-    return f"#{darkened_color_int:06x}"
-
-
-def collect_dim_reduced_embeds(
-    model_name, dim_reduced_embed_path, dim_reduction_model, **kwargs
-):
-    """
-    Return the dimensionality reduced embeddings of a model.
-
-    Parameters
-    ----------
-    model_name : str
-        name of model
-    dim_reduced_embed_path : pathlib.Path object
-        path to dim reduced embeddings
-    dim_reduction_model : str
-        name of feature extraction model
-
-    Returns
-    -------
-    dict
-        dimensionality reduced embeddings
-    """
-    files = list(dim_reduced_embed_path.iterdir())
-    if len(files) == 0:
-        logger.warning(
-            "No dimensionality reduced embeddings found for "
-            f"{dim_reduction_model}. In fact the directory "
-            f"{dim_reduced_embed_path} is empty. Deleting directory."
-        )
-        dim_reduced_embed_path.rmdir()
-        dim_reduced_embed_path = le.get_dim_reduc_path_func(
-            model_name, dim_reduction_model=dim_reduction_model, **kwargs
-        )
-        files = list(dim_reduced_embed_path.iterdir())
-    for file in files:
-        if file.suffix == ".json":  # and dim_reduction_model in file.stem:
-            with open(file, "r") as f:
-                embeds_dict = json.load(f)
-    return embeds_dict
-
-
-class EmbedAndLabelLoader:
-    def __init__(self, dim_reduction_model, dashboard=False, **kwargs):
-        self.labels = dict()
-        self.embeds = dict()
-        self.split_data = dict()
-        self.bool_noise = dict()
-        self.dashboard = dashboard
-        self.dim_reduction_model = dim_reduction_model
-        self.kwargs = kwargs
-
-    def get_data(self, model_name, label_by, remove_noise=False, **kwargs):
-        if not model_name in self.labels.keys():
-
-            tup = get_labels_for_plot(model_name, **self.kwargs)
-            self.labels[model_name], self.bool_noise[model_name] = tup
-
-            dim_reduced_embed_path = le.get_dim_reduc_path_func(
-                model_name, dim_reduction_model=self.dim_reduction_model, **kwargs
-            )
-
-            self.embeds[model_name] = collect_dim_reduced_embeds(
-                model_name, dim_reduced_embed_path, self.dim_reduction_model, **kwargs
-            )
-
-        if remove_noise:
-            return_labels = dict()
-            return_embeds = dict()
-            for key in self.labels[model_name].keys():
-                if "noise" in key:
-                    return_labels[key] = self.labels[model_name][key]
-                else:
-                    return_labels[key] = np.array(
-                        self.labels[model_name][key], dtype=object
-                    )[~self.bool_noise[model_name]]
-
-            return_embeds["x"] = np.array(self.embeds[model_name]["x"])[
-                ~self.bool_noise[model_name]
-            ]
-
-            return_embeds["y"] = np.array(self.embeds[model_name]["y"])[
-                ~self.bool_noise[model_name]
-            ]
-        else:
-            return_labels = self.labels[model_name]
-            return_embeds = self.embeds[model_name]
-
-        if label_by in return_labels:
-            return_splits = data_split_by_labels(return_embeds, return_labels[label_by])
-        else:
-            return [], [], {}
-        return (
-            return_labels[label_by],
-            return_embeds,
-            return_splits,
-        )
-
-
-def plot_embeddings(
-    loader,
-    model_name,
-    label_by,
-    paths=None,
-    dim_reduction_model=None,
-    axes=False,
-    fig=False,
-    dashboard=False,
-    dashboard_idx=None,
-    **kwargs,
-):
-    """
-    Generate figures and axes to plot points corresponding to embeddings.
-    This function can also be called and given figure and axes handeles.
-    In that case the existing handles will be used to add the points and
-    configure the axes and labels.
-
-    Parameters
-    ----------
-    loader : EmbedAndLabelLoader object
-        contains the labels and embeddings by model, for quicker loading
-    model_name : str
-        name of model
-    label_by : str, optional
-        key of default_labels dict, by default "audio_file_name"
-    paths : SimpleNamespace object, optional
-        object with path attributes, defaults to None
-    dim_reduction_model : str
-        name of dim reduced model
-    axes : plt object, optional
-        axes handle, by default False
-    fig : plt object, optional
-        figure handle, by default False
-    dashboard : bool, optional
-        whether the calls comes from the dashboard, by deafult False
-    dashboard_idx : int, optional
-        index of dashboard plot, relevant for legend placement
-
-    Returns
-    -------
-    plt object
-        axes handles is axes handles were given
-    dict
-        color dictionary for legend
-    list
-        plt point objects for legend of colorbar
-    """
-    labels, embeds, split_data = loader.get_data(model_name, label_by, **kwargs)
-
-    fig, axes, return_axes = init_embed_figure(fig, axes, **kwargs)
-    
-    if len(labels) == 0 and len(embeds) == 0:
-        return fig
-
-    if label_by == 'audio_file_name':
-        new_labels = [Path(l).stem+Path(l).suffix for l in labels]
-        new_split_data = dict()
-        for label in split_data.keys():
-            new_label = Path(label).stem+Path(label).suffix
-            new_split_data[new_label] = split_data[label]
-        split_data = new_split_data
-
-    c_label_dict = {lab: i for i, lab in enumerate(np.unique(labels))}
-    points = plot_embedding_points(
-        axes, embeds, split_data, labels, c_label_dict, **kwargs
-    )
-
-    if return_axes:
-        return axes, c_label_dict, points
-    elif dashboard:
-        if True:
-            # return plotly_mutual_information(
-            return SpectrogramPlot.plot_embeddings_px(
-                embeds, 
-                labels,
-                c_label_dict
-            )
-        else:
-            fig.set_size_inches(6, 5)
-            fig.set_dpi(300)
-            fig.tight_layout()
-            set_colorbar_or_legend(
-                fig,
-                axes,
-                points,
-                c_label_dict,
-                dashboard=dashboard,
-                label_by=label_by,
-                **kwargs,
-            )
-            return fig
-    else:
-        set_colorbar_or_legend(
-            fig, axes, points, c_label_dict, label_by=label_by, **kwargs
-        )
-
-        axes.set_title(f"{dim_reduction_model.upper()} embeddings")
-        fig.savefig(paths.plot_path.joinpath("embeddings.png"), dpi=300)
-        plt.close(fig)
-import pandas as pd
-import numpy as np
-import plotly.express as px
-import librosa as lb
-from scipy.signal.windows import tukey
-
-
-class SpectrogramPlot:
-    def __init__(self, audio_dir, **kwargs):
-        self.audio_dir = audio_dir
-        self.kwargs = kwargs
-
-    @staticmethod
-    def dummy_image():
-        # initial dummy figure, as a placeholder
-        fig = px.imshow(np.zeros((100, 100, 3), dtype=np.uint8))
-        fig.update_layout(
-            title="Click an embedding to see the corresponding spectrogram", 
-            margin=dict(l=20, r=20, t=40, b=20),
-            height=SPECTROGRAM_PLOT_HEIGHT,
-            xaxis={'visible': False}, 
-            yaxis={'visible': False}
-        )
-        return fig
-    
-    
-    def update_spectrogram(self, clickData=None, play_btn=None, autoplay_radio=None):
-        # Sohw black image initially
-        if not clickData:
-            return SpectrogramPlot.dummy_image()
-        
-        # Extract data from click
-        point_data = clickData.get('customdata', [None]*6)
-        audiofilename, start_s, end_s, idx, date, label_id = point_data
-        
-        # Load Audio
-        audio, sr, file_stem = self.load_audio(start_s, end_s, audiofilename)
-        spec_fig = self.create_specs(audio, sr)
-        
-        # Update title info
-        # heading = html.P([
-        #     f"file: {file_stem}", html.Br(),
-        #     f"time in file: {start_s}"
-        # ])
-
-        # # Handle Audio Playback
-        # # dash.ctx.triggered_id tells us exactly which input fired the callback
-        # trigger_id = dash.ctx.triggered_id
-        
-        # should_play = (autoplay_radio == "Autoplay on" 
-        #                and trigger_id == "bar_chart") or \
-        #                 (trigger_id == "play_audio_btn")
-
-        # if should_play:
-        #     self.play_audio(audio, sr)
-            
-        return spec_fig#, heading
-
-
-    # --- HELPERS ---
-    def create_specs(self, audio, audio_sr):
-        S = np.abs(lb.stft(audio, win_length=1024))
-        S_dB = lb.amplitude_to_db(S, ref=np.max)
-        f_max, S_dB = self.set_axis_lims_dep_sr(S_dB, audio_sr)
-        fig = px.imshow(
-            S_dB, origin='lower', aspect='auto',
-            y=np.linspace(0, f_max, S_dB.shape[0]),
-            x=np.linspace(0, 3, S_dB.shape[1]),
-            labels={'x': 'time (s)', 'y': 'freq (Hz)'}, 
-            color_continuous_scale='Viridis'
-        )
-        fig.update_layout(margin=dict(l=20, r=20, t=20, b=20))
-        return fig
-        
-    # def play_audio(self, audio, sr):
-    #     sd.play(self.fade_audio(audio), sr)
-        
-    def load_audio(self, start, end, filename):
-        path = Path(self.audio_dir) / filename
-        audio, sr = lb.load(path, offset=float(start), duration=float(end)-float(start))
-        if True:#self.use_tukey_filter:
-            audio = tukey(len(audio), alpha=0.01) * audio
-        audio_padded = lb.util.fix_length(
-            audio, size=int((end - start) * sr), mode=self.kwargs['padding']
-            )
-        return audio_padded, sr, path.stem
-    
-    def set_axis_lims_dep_sr(self, S_dB, audio_sr):
-        f_max = 48000 / 2
-        reduce = audio_sr / (f_max * 2)
-        S_dB = S_dB[:int(S_dB.shape[0] / reduce), :]
-        return f_max, S_dB
-
-
-    def plot_embeddings_px(
-        embeds,
-        labels,
-        c_label_dict,
-        label_by="label", # Added to use for titles/colorbar label
-        **kwargs
-    ):
-        # 1. Prepare Data
-        x_data = embeds['x']
-        y_data = embeds['y']
-        
-        audiofilenames = []
-        [
-            audiofilenames.extend([f] * nr) 
-            for f, nr in zip(
-                embeds['metadata']['audio_files']
-                , embeds['metadata']['nr_embeds_per_file']
-                )
-        ]
-        starts = embeds['timestamp']
-        ends = np.array(starts) + (
-            embeds['metadata']['segment_length (samples)'] 
-            / embeds['metadata']['sample_rate (Hz)']
-            )
-        ends = ends.tolist()
-        
-        # Calculate unique labels to decide on Legend vs Colorbar
-        unique_labels = np.unique(labels)
-        n_labels = len(unique_labels)
-        
-        # Create an integer mapping for high-cardinality plotting
-        # (Plotly needs numbers to generate a gradient colorbar)
-        label_to_id = {lbl: i for i, lbl in enumerate(unique_labels)}
-        label_ids = [label_to_id[l] for l in labels]
-
-        df = pd.DataFrame({
-            'x': x_data,
-            'y': y_data,
-            'label': labels,            # The actual string (for hover/legend)
-            'label_id': label_ids,      # The integer (for colorbar)
-            'audiofilename': audiofilenames,
-            'start': starts,
-            'end': ends,
-            'idx': range(len(audiofilenames))
-        })
-
-        # 2. Setup Figure based on Label Count
-        if n_labels > 20:
-            # --- HIGH CARDINALITY: Use Colorbar ---
-            # We map color to 'label_id' (int) to force a continuous scale
-            fig = px.scatter(
-                df, x='x', y='y',
-                color='label_id', 
-                hover_data={'label': True, 'label_id': False, 'audiofilename':True, 'start':True, 'end':True},
-                custom_data=['audiofilename', 'start', 'end', 'idx'],
-                title=f"Embedding Plot - {label_by}",
-                render_mode='webgl',
-                color_continuous_scale='Turbo' # High contrast for many classes
-            )
-
-            # Calculate Tick Positions (simulating your numpy logic)
-            # We want roughly 5-6 ticks evenly spaced
-            tick_vals = np.linspace(0, n_labels - 1, 6).astype(int)
-            tick_text = [unique_labels[i] for i in tick_vals]
-
-            # Customize the Colorbar to show Text instead of Numbers
-            fig.update_layout(
-                coloraxis_colorbar=dict(
-                    title=label_by,
-                    tickmode='array',
-                    tickvals=tick_vals,
-                    ticktext=tick_text,
-                    ticks='outside',
-                    len=1, # Height of colorbar (1 = 100%)
-                )
-            )
-        else:
-            # --- LOW CARDINALITY: Use Legend ---
-            # We map color to 'label' (string) to force a discrete legend
-            fig = px.scatter(
-                df, x='x', y='y',
-                color='label', 
-                hover_data=['filename', 'start', 'end'],
-                custom_data=['filename', 'start', 'end', 'idx'],
-                title=f"Embedding Plot - {label_by}",
-                render_mode='webgl'
-            )
-            
-            # Configure the Discrete Legend
-            fig.update_layout(
-                legend=dict(
-                    orientation="h", 
-                    yanchor="bottom", 
-                    y=1.02, 
-                    xanchor="right", 
-                    x=1,
-                    title_text="" # Hide legend title to save space
-                )
-            )
-
-        # 3. Global Layout adjustments (Applied to both modes)
-        fig.update_layout(
-            clickmode='event', 
-            # dragmode='lasso',
-            height=400,
-            hovermode='closest',
-            margin=dict(l=20, r=20, t=40, b=20),
-            # Ensure selection tools are available
-            modebar=dict(add=['lasso2d', 'select2d'], remove=['autoScale2d'])
-        )
-        
-        # Improve marker appearance
-        fig.update_traces(marker=dict(size=8, opacity=0.6))
-
-        return fig
-
-def init_embed_figure(fig, axes, bool_3d=False, **kwargs):
-    if not fig:
-        if bool_3d:
-            fig, axes = plt.subplots(subplot_kw={"projection": "3d"}, figsize=(12, 8))
-        else:
-            fig, axes = plt.subplots(figsize=(12, 8), dpi=400)
-        return_axes = False
-    else:
-        return_axes = True
-    axes.set_xticks([])
-    axes.set_yticks([])
-    return fig, axes, return_axes
-
-
-def get_labels_for_plot(model_name=None, **kwargs):
-    labels = dict()
-    labels = le.get_default_labels(model_name, **kwargs)
-
-    if le.get_paths(model_name).labels_path.joinpath("ground_truth.npy").exists():
-        ground_truth = le.get_ground_truth(model_name)
-        for label_column in [key for key in ground_truth.keys() if "label:" in key]:
-            label = label_column.split("label:")[-1]
-            inv = {v: k for k, v in ground_truth[f"label_dict:{label}"].items()}
-            inv[-1.0] = "noise"
-            inv[-2.0] = "noise"
-            # technically -2.0 is not noise, but corresponds to sections
-            # with multiple sources vocalizing simultaneously
-            if len(ground_truth[label_column].shape) > 1:
-                # TODO for display we're just taking the first label
-                labels[label] = [inv[v] for v in ground_truth[label_column][:, 0]]
-            else:
-                labels[label] = [inv[v] for v in ground_truth[label_column]]
-            bool_noise = np.array(labels[label]) == "noise"
-    else:
-        bool_noise = np.array([False] * len(list(labels.values())[0]))
-    if len(list(le.get_paths(model_name).clust_path.glob("*.npy"))) > 0:
-        clusts = [
-            np.load(f, allow_pickle=True).item()
-            for f in le.get_paths(model_name).clust_path.glob("*.npy")
-        ]
-        for clust in clusts:
-            for name, values in clust.items():
-                if "kmeans" in name:
-                    labels[name] = values
-                else:
-                    labels[name] = np.array(["noise"] * len(bool_noise), dtype=object)
-                    labels[name][~bool_noise] = [inv[v] for v in values]
-
-    return labels, bool_noise
-
-
-def set_colorbar_or_legend(fig, axes, points, c_label_dict, label_by, **kwargs):
-    if len(c_label_dict.keys()) > 20:
-        if isinstance(list(c_label_dict.keys())[0], int):
-            fontsize = 9
-        elif isinstance(list(c_label_dict.keys())[0], np.int32):
-            fontsize = 9
-        elif len(list(c_label_dict.keys())[0]) < 12:
-            fontsize = 9
-        else:
-            fontsize = 6
-
-        # Shrink main plot area to make space for colorbar
-        fig.subplots_adjust(right=0.7)
-
-        # Add colorbar axis manually (x0, y0, width, height) in figure coords
-        cbar_ax = fig.add_axes([0.72, 0.05, 0.03, 0.9])  # tweak as needed
-
-        # Create colorbar in the custom axis
-        cbar = fig.colorbar(points, cax=cbar_ax)
-
-        locs = [*(int(len(c_label_dict) / 5) * np.arange(5)), -1]
-        cbar.set_ticks([list(c_label_dict.values())[loc] for loc in locs])
-        cbar.set_ticklabels(
-            [list(c_label_dict.keys())[loc] for loc in locs], fontsize=fontsize
-        )
-        cbar.set_label(label_by.replace("_", " "), fontsize=10)
-    else:
-        hands, labs = axes.get_legend_handles_labels()
-        fig, axes = set_legend(hands, labs, fig, axes, **kwargs)
-    return fig, axes
-
-
-def plot_embedding_points(
-    axes, embeds, split_data, labels, c_label_dict, remove_noise=False, **kwargs
-):
-    """
-    Plot embeddings in scatter plot.
-
-    Parameters
-    ----------
-    axes : plt object
-        axes handle
-    embeds : dict
-        embeddings
-    split_data : dict
-        data split by label
-    labels : list
-        labels of the data
-    c_label_dict : dict
-        linking labels to ints for coloring
-    remove_noise : bool, optional
-        remove noise or not, defaults to False
-
-    Returns
-    -------
-    plt object
-        axes points
-    """
-    if len(c_label_dict.keys()) > 20:
-        import matplotlib.cm as cm
-
-        cmap = cm.viridis  # or 'plasma', 'inferno', 'magma', etc.
-        # if remove_noise:
-        #     bool_labels = np.array(labels) != "noise"
-        #     labels = np.array(labels)[bool_labels]
-        # else:
-        #     bool_labels = [True] * len(labels)
-
-        num_labels = np.array([c_label_dict[lab] for lab in labels])
-        if not len(labels) == len(embeds['x']):
-            raise AssertionError(
-                f"The number of labels is {len(labels)} whereas the number of "
-                f"embedding points is {len(embeds['x'])}. This mismatch could "
-                "be the result of an incomplete run and bacpipe is using "
-                "the dim_reduced_embeddings corresponding to that. Check if in your results folder "
-                "there are not multiple dim_reduced_embeddings, and if so, delete the incomplete one."
-            )
-        points = axes.scatter(
-            # np.array(embeds["x"])[bool_labels],
-            # np.array(embeds["y"])[bool_labels],
-            np.array(embeds["x"]),
-            np.array(embeds["y"]),
-            c=num_labels,
-            label=labels,
-            s=1,
-            cmap=cmap,
-        )
-    else:
-        cmap = plt.cm.tab20
-        colors = cmap(np.arange(len(c_label_dict.keys())) % cmap.N)
-        for idx, (label, data) in enumerate(split_data.items()):
-            if remove_noise and label == "noise":
-                continue
-            points = axes.scatter(
-                data[0],
-                data[1],
-                label=label,
-                s=1,
-                color=colors[idx],
-            )
-    return points
-
-
-def set_legend(
-    handles, labels, fig, axes, bool_plot_centroids=False, dashboard=False, **kwargs
-):
-    """
-    Create the legend for embeddings visualization plots.
-
-    Parameters
-    ----------
-    handles : list
-        list of legend handles
-    labels : list
-        list of labels for legend
-    fig : plt.fig object
-        figure handle
-    axes : plt.axes object
-        axes handle
-    bool_plot_centroids : bool, optional
-        if True centroids of each class will be plotted, by default True
-    dashboard : bool
-        if dashboard called this function or not
-
-    Returns
-    -------
-    plt.fig object
-        figure handle
-    plt.axes object
-        axes handle
-    """
-
-    # Calculate number of columns dynamically based on the number of labels
-    num_labels = len(labels)  # Number of labels in the legend
-    ncol = min(num_labels, 5)  # Use 6 columns or fewer if there are fewer labels
-
-    if bool_plot_centroids:
-        custom_marker = plt.scatter(
-            [], [], marker="x", color="black", s=10
-        )  # Empty scatter, only for the legend
-        new_handles = handles[::2] + [custom_marker]
-        new_labels = labels[::2] + ["centroids"]
-    else:
-        new_handles = handles
-        new_labels = labels
-    if dashboard:
-        fig.subplots_adjust(right=0.7)
-
-        fig.legend(
-            new_handles,
-            new_labels,
-            loc="outside right",
-            markerscale=4 if dashboard else 6,
-            fontsize=7,
-            frameon=False,
-        )
-    else:
-
-        fig.subplots_adjust(bottom=0.2)
-        fig.legend(
-            new_handles,
-            new_labels,  # Use the handles and labels from the plot
-            loc="outside lower center",  # Center the legend
-            ncol=ncol,  # Number of columns
-            markerscale=6,
-        )
-    return fig, axes
-
-
-def data_split_by_labels(embeds_dict, labels):
-    """
-    Split data by labels for scatterplots.
-
-    Parameters
-    ----------
-    embeds_dict : dict
-        embeddings by model
-    labels : list
-        list of labels
-
-    Returns
-    -------
-    dict
-        x and y data corresponding to labels
-    """
-    split_data = {}
-    uni_labels = np.unique(labels)
-    if len(uni_labels) > 20:
-        split_data["all"] = np.array(
-            [
-                np.array(embeds_dict["x"]),
-                np.array(embeds_dict["y"]),
-            ]
-        )
-    else:
-        for label in uni_labels:  # TODO don't do this for more than 20 categories
-            split_data[str(label)] = np.array(
-                [
-                    np.array(embeds_dict["x"])[np.array(labels) == label],
-                    np.array(embeds_dict["y"])[np.array(labels) == label],
-                ]
-            )
-
-    return split_data
-
-
-def return_rows_cols(num):
-    if num <= 3:
-        return 1, 3
-    elif num > 3 and num <= 6:
-        return 2, 3
-    elif num > 6 and num <= 9:
-        return 3, 3
-    elif num > 9 and num <= 12:
-        return 3, 4
-    elif num > 12 and num <= 16:
-        return 4, 4
-    elif num > 16 and num <= 20:
-        return 4, 5
-    else:
-        return 5, num // 5
-
-
-def set_figsize_for_comparison(rows, cols):
-    if rows == 1:
-        return (12, 5)
-    elif rows == 2:
-        return (12, 7)
-    elif rows == 3:
-        return (12, 8)
-    elif rows > 3:
-        return (12, 10)
-
-
-def plot_comparison(
-    plot_path,
-    models,
-    dim_reduction_model,
-    bool_spherical=False,
-    dashboard=False,
-    loader=None,
-    evaluation_task=[],
-    **kwargs,
-):
-    """
-    Create big overview visualization of all embeddings spaces. Labels
-    are chosen from ground_truth and if that does not exist, default
-    lables are used.
-
-    Parameters
-    ----------
-    plot_path : pathlib.Path object
-        path to store overview plots
-    models : list
-        list of models
-    dim_reduction_model : str
-        name of dimensionality reduction model
-    bool_spherical : bool, optional
-        if True 3d embeddings will be plotted, by default False
-    dashboard : bool, optional
-        if dashboard called this function or not
-    loader : EmbedAndLabelLoader object
-        object containing embeds and labels by model for quicker loading
-    evaluation_task : list, optional
-        list of tasks to evaluate, by default []
-
-    Returns
-    -------
-    plt object
-        figure handle
-    """
-    rows, cols = return_rows_cols(len(models))
-
-    if not bool_spherical:
-        fig = Figure(figsize=set_figsize_for_comparison(rows, cols))
-        axes = fig.subplots(rows, cols)
-    else:
-        fig = Figure(figsize=set_figsize_for_comparison(rows, cols))
-        axes = fig.subplots(
-            rows,
-            cols,
-            subplot_kw={"projection": "3d"},
-        )
-    if not dashboard:
-        vis_loader = EmbedAndLabelLoader(dim_reduction_model, **kwargs)
-    else:
-        vis_loader = loader
-
-    c_label_dict, points = {}, {}
-    for idx, model in enumerate(models):
-        paths = le.get_paths(model)
-
-        axes.flatten()[idx], c_label_dict[idx], points[idx] = plot_embeddings(
-            vis_loader,
-            model,
-            paths=paths,
-            dim_reduction_model=dim_reduction_model,
-            axes=axes.flatten()[idx],
-            fig=fig,
-            bool_plot_centroids=False,
-            dashboard=dashboard,
-            **kwargs,
-        )
-        axes.flatten()[idx].set_title(f"{model.upper()}")
-
-    fig.tight_layout()
-    fig.subplots_adjust(top=0.9, bottom=0.2)
-    colorbar_idx = np.argmax([len(d) for d in c_label_dict.values()])
-
-    fig, _ = set_colorbar_or_legend(
-        fig,
-        axes.flatten()[colorbar_idx],
-        points[colorbar_idx],
-        c_label_dict[colorbar_idx],
-        dashboard=dashboard,
-        **kwargs,
-    )
-    [ax.remove() for ax in axes.flatten()[idx + 1 :]]
-    if "clustering" in evaluation_task:
-        reorder_embeddings_by_clustering_performance(plot_path, axes, models)
-
-    fig.suptitle(f"Comparison of {dim_reduction_model} embeddings", fontweight="bold")
-    if not dashboard:
-        fig.savefig(plot_path.joinpath("comp_fig.png"), dpi=300)
-        plt.close(fig)
-    else:
-        return fig
-
-
-def reorder_embeddings_by_clustering_performance(
-    plot_path, axes, models, order_metric="ground_truth-kmeans"
-):
-    """
-    Reorder the embedding overview plot by clustering performance.
-
-    Parameters
-    ----------
-    plot_path : pathlib.Path object
-        path to store plots and results comparing all models
-    axes : plt.axes object
-        handle for figures axes
-    models : list
-        list of models
-    order_metric : str
-        key corresponding to a metric in the clustering_results.json file.
-        Defaults to "ARI(kmeans)"
-    """
-    clust_dict = json.load(open(plot_path.joinpath("clustering_results.json"), "r"))
-    new_order = dict(
-        sorted(
-            clust_dict.items(), key=lambda kv: kv[1]["ARI"][order_metric], reverse=True
-        )
-    )
-    positions = {mod: ax.get_position() for mod, ax in zip(new_order, axes.flatten())}
-    for model, ax in zip(models, axes.flatten()):
-        if not model in positions.keys():
-            continue
-        ax.set_position(positions[model])
-
-
-def plot_classification_results(
-    task_name,
-    paths=None,
-    metrics=None,
-    return_fig=False,
-    path_func=None,
-    model_name=None,
-):
-    """
-    Save model specific classification results in the model specific
-    plot path, displayed as horizontal bars.
-
-    Parameters
-    ----------
-    task_name : str
-        name of task
-    paths : SimpleNamespace object
-        path to store plots
-    metrics : dict
-        classification performance
-    return_fig : bool
-        if True the figure will be returned, by default False
-    path_func : function
-        function to return the paths when model name is given
-    model_name : str
-        name of model, by default None
-
-    Returns
-    -------
-    plt object
-        figure handle
-    """
-    if path_func and model_name:
-        paths = path_func(model_name)
-    if not metrics:
-        class_path = paths.class_path / f"class_results_{task_name}.json"
-        if not class_path.exists():
-            error = (
-                f"\nThe classification file {class_path} does not exist. Perhaps it was not "
-                "created yet. To avoid getting this error, make sure you have not "
-                " included 'classification' in the 'evaluation_tasks'. If you want to compute "
-                "classification, make sure to set `overwrite=True`."
-            )
-            logger.exception(error)
-            raise AssertionError(error)
-
-        with open(paths.class_path / f"class_results_{task_name}.json", "r") as f:
-            metrics = json.load(f)
-
-    # Filter overall metrics if needed
-    metrics["overall"] = {
-        k: v for k, v in metrics["overall"].items() if not "micro" in k
-    }
-
-    # Sort classes by accuracy for better visualization
-    class_items = sorted(
-        metrics["per_class_accuracy"].items(), key=lambda x: x[1], reverse=True
-    )
-    class_names = [item[0] for item in class_items]
-    class_values = [item[1] for item in class_items]
-
-    # Set figure size based on number of classes and return_fig
-    if return_fig:
-        # For dashboard, make height adapt to number of classes
-        height = max(4, len(class_names) * 0.3)
-        fig, ax = plt.subplots(1, 1, figsize=(5, height))
-        fontsize = 10
-    else:
-        height = max(8, len(class_names) * 0.4)
-        fig, ax = plt.subplots(1, 1, figsize=(12, height))
-        fontsize = 14
-
-    model_name = paths.labels_path.parent.stem
-    cmap = plt.cm.tab10
-    colors = cmap(np.arange(len(class_names)) % cmap.N)
-
-    # Create horizontal bars
-    ax.barh(
-        range(len(class_names)),
-        class_values,
-        height=0.6,
-        color=colors,
-    )
-
-    # Create metrics string
-    metrics_string = "".join(
-        [f"{k}: {v:.3f} | " for k, v in metrics["overall"].items()]
-    )
-
-    fig.suptitle(
-        f"Classwise accuracy for {task_name} "
-        f"classification with {model_name.upper()} embeddings\n"
-        f"{metrics_string}",
-        fontsize=fontsize,
-    )
-
-    # Adjust labels for horizontal orientation
-    ax.set_xlabel("Accuracy")
-    ax.set_ylabel("Classes")
-    ax.set_yticks(range(len(class_names)))
-    ax.set_yticklabels(class_names, fontsize=8)
-
-    # Add value labels at the end of each bar
-    for i, v in enumerate(class_values):
-        ax.text(v + 0.01, i, f"{v:.2f}", va="center", fontsize=8)
-
-    # Set x-axis limits for better visualization
-    ax.set_xlim(0, min(1.0, max(class_values) * 1.15))
-
-    # Add grid lines for easier reading
-    ax.grid(axis="x", linestyle="--", alpha=0.7)
-
-    # Adjust layout
-    fig.tight_layout()
-    fig.subplots_adjust(top=0.9)
-
-    if return_fig:
-        return fig
-
-    path = paths.plot_path
-    fig.savefig(
-        path.joinpath(f"class_results_{task_name}_{model_name}.png"),
-        dpi=300,
-    )
-    plt.close(fig)
-
-
-def load_results(path_func, task, model_list):
-    """
-    Load the task results into a dict and return them. For classification
-    multiple subtasks exist, so do them seperately.
-
-    Parameters
-    ----------
-    path_func : function
-        returns model specific tasks when model is given
-    task : str
-        name of task
-    model_list : list
-        list of models
-
-    Returns
-    -------
-    dict
-        performance for different tasks and models
-    """
-    metrics = {}
-    for model_name in model_list:
-        paths = path_func(model_name)
-        for file in getattr(paths, f"{task[:5]}_path").rglob("*results*.json"):
-            if task == "classification":
-                subtask = file.stem.split("_")[-1]
-                metrics[f"{model_name}({subtask})"] = json.load(open(file, "r"))
-            else:
-                metrics[model_name] = json.load(open(file, "r"))
-    return metrics
 
 
 def visualise_results_across_models(plot_path, task_name, model_list):
@@ -1357,132 +359,3 @@ def plot_overview_metrics(
     )
     plt.close(fig)
 
-
-def plot_per_class_metrics(plot_path, task_name, model_list, metrics):
-    """
-    Visualization of per class results. Resulting figure is stored in
-    plot path. Models are sorted by the value of the first entry.
-
-    Parameters
-    ----------
-    plot_path : pathlib.Path object
-        path to store plot in
-    task_name : str
-        name of task
-    model_list : list
-        list of models
-    metrics : dict
-        performance dictionary
-    """
-    per_class_metrics = {m: v["per_class_accuracy"] for m, v in metrics.items()}
-    overall_metrics = {m: v["overall"] for m, v in metrics.items()}
-    num_classes = len(per_class_metrics[model_list[0]].keys())
-    fig_width = max(12, num_classes * 0.5)
-    fig, ax = plt.subplots(1, 1, figsize=(fig_width, 8))
-
-    cmap = plt.cm.tab10
-    model_colors = cmap(np.arange(len(model_list)) % cmap.N)
-
-    d = {m: v["macro_accuracy"] for m, v in overall_metrics.items()}
-    model_list = sorted(d, key=d.get, reverse=True)
-    all_classes = sorted(per_class_metrics[model_list[0]].keys())
-
-    for i, model_name in enumerate(model_list):
-        class_values = per_class_metrics[model_name].values()
-
-        ax.scatter(
-            np.arange(len(class_values)),
-            class_values,
-            color=model_colors[i],
-            label=f"{model_name.upper()} "
-            + f"(accuracy: {overall_metrics[model_name]['macro_accuracy']:.3f})",
-            s=100,
-        )
-
-        ax.plot(
-            np.arange(len(class_values)),
-            class_values,
-            color=model_colors[i],
-            linestyle="-",  # Solid line
-            linewidth=1.5,
-        )
-
-    fig.suptitle(
-        f"Per class metrics for {task_name} across models",
-        fontsize=14,
-    )
-    ax.set_ylabel("Accuracy")
-    ax.set_xlabel("Classes")
-    ax.set_xticks(np.arange(len(all_classes)))
-    ax.set_xticklabels(all_classes, rotation=90)
-
-    ax.legend(loc="upper left", bbox_to_anchor=(1.05, 1), title="Models", fontsize=10)
-
-    fig.subplots_adjust(right=0.65, bottom=0.3)
-    file_name = (
-        f"comparison_{task_name.replace(' ', '_')}_" 
-        + "-".join([m[:2] for m in model_list]) 
-        + ".png"
-    )
-    plot_path.mkdir(exist_ok=True, parents=True)
-    fig.savefig(
-        plot_path.joinpath(file_name),
-        dpi=300,
-    )
-    plt.close(fig)
-
-        
-def plot_classification_heatmap(
-    accumulated_presence, timestamps, accumulate_by, species, threshold
-    ):
-    
-    fig = Figure(figsize=[11, 8])
-    ax = fig.subplots()
-    
-    fig.suptitle(
-        f'Presence heatmap for {species} with threshold of {threshold}',
-        fontsize=10
-        )
-    sns.heatmap(
-        accumulated_presence.T, 
-        vmin=0,
-        # vmax=1,
-        cmap='viridis',
-        cbar_kws={'label': 'Binary presence per hour'},
-        ax=ax
-        )
-    
-    y_locs, yticklabels = ax.get_yticks(), ax.get_yticklabels()
-    if accumulate_by == 'day':
-        labels = np.unique([ts.date() for ts in timestamps])
-        ax.set_ylabel('dates')
-    elif accumulate_by == 'month':
-        labels = np.unique(
-            [f'{date.year}-{date.month}' for date in timestamps], 
-            axis=0
-            )
-        ax.set_ylabel('months')
-    elif accumulate_by == 'week':
-        labels = np.unique(
-            [f'{date.year}-{date.isocalendar().week}' for date in timestamps], 
-            axis=0
-            )
-        ax.set_ylabel('weeks')
-    selected_labels = labels[[int(i.get_text()) for i in yticklabels]]
-    x_locs, labels = ax.get_xticks(), ax.get_xticklabels()
-    x_idxs = [0, 6, 12, 18, 23]
-    ax.set_xticks(x_locs[x_idxs], np.array(labels)[x_idxs])
-        
-    
-    ax.set_xlabel('hours')
-    ax.set_yticks(y_locs)
-    ax.set_yticklabels(selected_labels)
-    
-    # force the rotation on the axis itself
-    ax.tick_params(axis='y', rotation=0)
-    
-    
-    fig.set_size_inches(6, 5)
-    fig.set_dpi(300)
-    fig.tight_layout()
-    return fig
